+++ lucene/backward-codecs/src/java/org/apache/lucene/backward_codecs/lucene92/Lucene92Codec.java
@@ -154,7 +154,7 @@ public class Lucene92Codec extends Codec {
   }
 
   @Override
-  public CompoundFormat compoundFormat() {
+  public final CompoundFormat compoundFormat() {
     return compoundFormat;
   }
