+++ lucene/backward-codecs/src/java/org/apache/lucene/backward_codecs/lucene94/Lucene94Codec.java
@@ -154,7 +154,7 @@ public class Lucene94Codec extends Codec {
   }
 
   @Override
-  public CompoundFormat compoundFormat() {
+  public final CompoundFormat compoundFormat() {
     return compoundFormat;
   }
