+++ lucene/backward-codecs/src/resources/META-INF/services/org.apache.lucene.codecs.Codec
@@ -22,4 +22,3 @@ org.apache.lucene.backward_codecs.lucene91.Lucene91Codec
 org.apache.lucene.backward_codecs.lucene92.Lucene92Codec
 org.apache.lucene.backward_codecs.lucene94.Lucene94Codec
 org.apache.lucene.backward_codecs.lucene95.Lucene95Codec
-org.apache.lucene.backward_codecs.lucene99.Lucene99Codec
