+++ lucene/backward-codecs/src/test/org/apache/lucene/backward_codecs/lucene91/Lucene91RWCodec.java
@@ -16,10 +16,8 @@
  */
 package org.apache.lucene.backward_codecs.lucene91;
 
-import org.apache.lucene.backward_codecs.lucene90.Lucene90RWCompoundFormat;
 import org.apache.lucene.backward_codecs.lucene90.Lucene90RWPostingsFormat;
 import org.apache.lucene.backward_codecs.lucene90.Lucene90RWSegmentInfoFormat;
-import org.apache.lucene.codecs.CompoundFormat;
 import org.apache.lucene.codecs.KnnVectorsFormat;
 import org.apache.lucene.codecs.PostingsFormat;
 import org.apache.lucene.codecs.SegmentInfoFormat;
@@ -27,7 +25,6 @@ import org.apache.lucene.codecs.perfield.PerFieldKnnVectorsFormat;
 
 public class Lucene91RWCodec extends Lucene91Codec {
 
-  private final CompoundFormat compoundFormat = new Lucene90RWCompoundFormat();
   private final KnnVectorsFormat defaultKnnVectorsFormat;
   private final KnnVectorsFormat knnVectorsFormat =
       new PerFieldKnnVectorsFormat() {
@@ -44,11 +41,6 @@ public class Lucene91RWCodec extends Lucene91Codec {
         new Lucene91RWHnswVectorsFormat(
             Lucene91HnswVectorsFormat.DEFAULT_MAX_CONN,
             Lucene91HnswVectorsFormat.DEFAULT_BEAM_WIDTH);
-  }
-
-  @Override
-  public CompoundFormat compoundFormat() {
-    return compoundFormat;
   }
 
   @Override
