+++ lucene/backward-codecs/src/test/org/apache/lucene/backward_codecs/lucene92/Lucene92RWCodec.java
@@ -16,10 +16,8 @@
  */
 package org.apache.lucene.backward_codecs.lucene92;
 
-import org.apache.lucene.backward_codecs.lucene90.Lucene90RWCompoundFormat;
 import org.apache.lucene.backward_codecs.lucene90.Lucene90RWPostingsFormat;
 import org.apache.lucene.backward_codecs.lucene90.Lucene90RWSegmentInfoFormat;
-import org.apache.lucene.codecs.CompoundFormat;
 import org.apache.lucene.codecs.KnnVectorsFormat;
 import org.apache.lucene.codecs.PostingsFormat;
 import org.apache.lucene.codecs.SegmentInfoFormat;
@@ -28,7 +26,6 @@ import org.apache.lucene.codecs.perfield.PerFieldKnnVectorsFormat;
 /** Implements the Lucene 9.2 index format for backwards compat testing */
 public class Lucene92RWCodec extends Lucene92Codec {
 
-  private final CompoundFormat compoundFormat = new Lucene90RWCompoundFormat();
   private final KnnVectorsFormat defaultKnnVectorsFormat;
   private final KnnVectorsFormat knnVectorsFormat =
       new PerFieldKnnVectorsFormat() {
@@ -46,11 +43,6 @@ public class Lucene92RWCodec extends Lucene92Codec {
         new Lucene92RWHnswVectorsFormat(
             Lucene92HnswVectorsFormat.DEFAULT_MAX_CONN,
             Lucene92HnswVectorsFormat.DEFAULT_BEAM_WIDTH);
-  }
-
-  @Override
-  public CompoundFormat compoundFormat() {
-    return compoundFormat;
   }
 
   @Override
