+++ lucene/backward-codecs/src/test/org/apache/lucene/backward_codecs/lucene95/Lucene95RWCodec.java
@@ -16,11 +16,9 @@
  */
 package org.apache.lucene.backward_codecs.lucene95;
 
-import org.apache.lucene.backward_codecs.lucene90.Lucene90RWCompoundFormat;
 import org.apache.lucene.backward_codecs.lucene90.Lucene90RWPostingsFormat;
 import org.apache.lucene.backward_codecs.lucene90.Lucene90RWSegmentInfoFormat;
 import org.apache.lucene.backward_codecs.lucene94.Lucene94HnswVectorsFormat;
-import org.apache.lucene.codecs.CompoundFormat;
 import org.apache.lucene.codecs.KnnVectorsFormat;
 import org.apache.lucene.codecs.PostingsFormat;
 import org.apache.lucene.codecs.SegmentInfoFormat;
@@ -29,7 +27,6 @@ import org.apache.lucene.codecs.perfield.PerFieldKnnVectorsFormat;
 /** Implements the Lucene 9.5 index format for backwards compat testing */
 public class Lucene95RWCodec extends Lucene95Codec {
 
-  private final CompoundFormat compoundFormat = new Lucene90RWCompoundFormat();
   private final KnnVectorsFormat defaultKnnVectorsFormat;
   private final KnnVectorsFormat knnVectorsFormat =
       new PerFieldKnnVectorsFormat() {
@@ -47,11 +44,6 @@ public class Lucene95RWCodec extends Lucene95Codec {
         new Lucene95RWHnswVectorsFormat(
             Lucene94HnswVectorsFormat.DEFAULT_MAX_CONN,
             Lucene94HnswVectorsFormat.DEFAULT_BEAM_WIDTH);
-  }
-
-  @Override
-  public CompoundFormat compoundFormat() {
-    return compoundFormat;
   }
 
   @Override
