+++ lucene/core/src/resources/META-INF/services/org.apache.lucene.codecs.Codec
@@ -13,4 +13,4 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-org.apache.lucene.codecs.lucene100.Lucene100Codec
+org.apache.lucene.codecs.lucene99.Lucene99Codec
